Add initial code for airfoil simulation
Adds a test case for 2d airfoil simulation and a class for airfoils.
The parametric airfoil class can load a YAML file containing airfoil parameters.
This depends on the Yams YAML package.
For conditions outside the range of the airfoil model, we simply copy the first and last entry in the coefficiients.

The YAML format can support multiple definitions for different Re numbers.
But the loader and class only cares about the first one for now.
The idea being that we should be able to interpolate based on Re at some point in the future.